Add database tests and fix errors that cropped up (#669)

+++ app/db/impl/model/common/Downloadable.scala
@@ -16,6 +16,6 @@ trait Downloadable extends Model { self =>
     *
     * @return Downloads model has
     */
-  def downloadCount: Int
+  def downloadCount: Long
 
 }
